submission/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
passlib==1.7.4
email-validator==2.2.0
pytest==8.4.1
httpx==0.28.1
